test(testcase): updating test cases and snapshot
