+++ packages/react/src/components/CardEditor/CardEditor.test.jsx
@@ -1,10 +1,31 @@
 import React from 'react';
-import { render, screen } from '@testing-library/react';
+import { render, screen, fireEvent } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom/extend-expect';
 import { EscalatorDown } from '@carbon/pictograms-react';
 
-import CardEditor from './CardEditor';
+import CardEditor, { handleSubmit, hideCardPropertiesForEditor } from './CardEditor';
+
+const mockSetError = jest.fn();
+const mockOnChange = jest.fn();
+const mockOnCardJsonPreview = jest.fn();
+const mockSetShowEditor = jest.fn();
+const mockOnValidateCardJson = jest.fn().mockImplementation(() => []);
+
+const cardConfig = {
+  title: 'timeSeries',
+  size: 'MEDIUM',
+  type: 'TIMESERIES',
+  content: {
+    series: [
+      {
+        dataItemId: 'torque',
+        dataSourceId: 'torque_id',
+        label: 'Torque',
+      },
+    ],
+  },
+};
 
 describe('CardEditor', () => {
   const actions = {
@@ -19,6 +40,22 @@ describe('CardEditor', () => {
     size: 'SMALL',
     type: 'VALUE',
   };
+
+  it('should fire onCardJsonPreview when opening the json modal', () => {
+    render(
+      <CardEditor
+        cardConfig={cardConfig}
+        onChange={mockOnChange}
+        onCardJsonPreview={mockOnCardJsonPreview}
+      />
+    );
+    const openJsonBtn = screen.getByText('Open JSON editor');
+    expect(openJsonBtn).toBeInTheDocument();
+
+    fireEvent.click(openJsonBtn);
+
+    expect(mockOnCardJsonPreview).toHaveBeenCalledWith(cardConfig);
+  });
 
   it('is selectable by testID and testId', () => {
     jest.spyOn(console, 'error').mockImplementation(() => {});
@@ -220,6 +257,193 @@ describe('CardEditor', () => {
       timeRange: 'thisWeek',
       title: 'New card',
       type: 'VALUE',
+    });
+  });
+
+  // would like to do react-testing-library tests with this, but we're unable to render the actual editor,
+  // meaning we can't fire user events on the form
+  describe('handleSubmit', () => {
+    it('should throw error if JSON is empty', () => {
+      handleSubmit('', '', mockSetError, mockOnValidateCardJson, mockOnChange, mockSetShowEditor);
+      expect(mockSetError).toBeCalledWith('Unexpected end of JSON input');
+    });
+    it('should call onChange and setShowEditor if JSON is valid', () => {
+      handleSubmit('{}', '', mockSetError, mockOnValidateCardJson, mockOnChange, mockSetShowEditor);
+      expect(mockOnChange).toBeCalled();
+      expect(mockSetShowEditor).toBeCalledWith(false);
+    });
+    it('should call onChange with content section changese', () => {
+      handleSubmit(
+        '{"content":"my content"}',
+        '',
+        mockSetError,
+        mockOnValidateCardJson,
+        mockOnChange,
+        mockSetShowEditor
+      );
+      expect(mockOnChange).toBeCalledWith(expect.objectContaining({ content: 'my content' }));
+    });
+    it('should throw error if JSON is not valid', () => {
+      handleSubmit(
+        '1234',
+        '',
+        mockSetError,
+        mockOnValidateCardJson,
+        mockOnChange,
+        mockSetShowEditor
+      );
+      expect(mockSetError).toBeCalledWith('1234 is not valid JSON');
+    });
+  });
+  describe('hideCardPropertiesForEditor', () => {
+    it('should hide properties in the attributes section of a card', () => {
+      const sanitizedCard = hideCardPropertiesForEditor({
+        type: 'VALUE',
+        content: {
+          attributes: [
+            {
+              aggregationMethods: [],
+              aggregationMethod: '',
+              grain: '',
+              dataSourceId: 'torque',
+              label: 'Torque',
+            },
+          ],
+        },
+      });
+      expect(sanitizedCard).toEqual({
+        type: 'VALUE',
+        content: {
+          attributes: [
+            {
+              aggregationMethod: '',
+              dataSourceId: 'torque',
+              label: 'Torque',
+            },
+          ],
+        },
+      });
+    });
+    it('should hide properties in the series section of a card', () => {
+      const sanitizedCard = hideCardPropertiesForEditor({
+        type: 'TIMESERIES',
+        content: {
+          series: [
+            {
+              aggregationMethods: [],
+              aggregationMethod: '',
+              grain: '',
+              dataSourceId: 'torque',
+              label: 'Torque',
+            },
+          ],
+        },
+      });
+      expect(sanitizedCard).toEqual({
+        type: 'TIMESERIES',
+        content: {
+          series: [
+            {
+              aggregationMethod: '',
+              dataSourceId: 'torque',
+              label: 'Torque',
+            },
+          ],
+        },
+      });
+    });
+    it('should hide properties in the columns section of a card', () => {
+      const sanitizedCard = hideCardPropertiesForEditor({
+        type: 'TABLE',
+        content: {
+          columns: [
+            {
+              aggregationMethods: [],
+              aggregationMethod: '',
+              grain: '',
+              dataSourceId: 'torque',
+              label: 'Torque',
+            },
+          ],
+        },
+      });
+      expect(sanitizedCard).toEqual({
+        type: 'TABLE',
+        content: {
+          columns: [
+            {
+              aggregationMethod: '',
+              dataSourceId: 'torque',
+              label: 'Torque',
+            },
+          ],
+        },
+      });
+    });
+    it('should hide properties in the hotspots section of a card', () => {
+      const sanitizedCard = hideCardPropertiesForEditor({
+        type: 'IMAGE',
+        values: {
+          hotspots: [
+            {
+              x: 35,
+              y: 65,
+              icon: 'InformationFilled24',
+              color: 'green',
+              content: {
+                title: 'My Device',
+                description: 'Description',
+                attributes: [
+                  {
+                    dataItemId: 'temperature',
+                    dataSourceId: 'temperature',
+                    grain: '',
+                    aggregationMethods: [],
+                    label: 'Temp',
+                    precision: 2,
+                  },
+                ],
+              },
+            },
+          ],
+        },
+      });
+      expect(sanitizedCard).toEqual({
+        type: 'IMAGE',
+        values: {
+          hotspots: [
+            {
+              x: 35,
+              y: 65,
+              icon: 'InformationFilled24',
+              color: 'green',
+              content: {
+                title: 'My Device',
+                description: 'Description',
+                attributes: [
+                  {
+                    dataItemId: 'temperature',
+                    dataSourceId: 'temperature',
+                    label: 'Temp',
+                    precision: 2,
+                  },
+                ],
+              },
+            },
+          ],
+        },
+      });
+    });
+    it('should hide the content for a custom card', () => {
+      const sanitizedCard = hideCardPropertiesForEditor({
+        type: 'MY_CUSTOM_TYPE',
+        title: 'myCustomCard',
+        content: 'Custom card content',
+      });
+      expect(sanitizedCard).toEqual({
+        type: 'MY_CUSTOM_TYPE',
+        title: 'myCustomCard',
+      });
     });
   });
 });
